Refactor : Storage, Profile, FollowInfo 리팩터링 (#55)
* Refactor : 파일 위치 변경, 주석 삭제, 경로 변경

* Refactor : API 수정

* Refactor : Tab 공통 컴포넌트

* Refactor : Storage 컴포넌트 분리

* Refactor : play에 object fit

* Refactor : 컴포넌트, 훅 기반 FollowInfo 수정

* Refactor : Profile 컴포넌트 분리 및 적용

* Refactor : QueryKey Constant 수정

* Refactor : 안 쓰는 api 주석처리, isError, isLoading 추가

* Refactor : 위치 변경
