+++ src/api/useFollowList.ts
@@ -1,23 +1,22 @@
 import { useQuery } from "@tanstack/react-query";
-import axiosInstance from "@/api/axiosInstance";
+import { fetchFollowList } from "@/db/follow";
+import { QUERY_KEYS } from "@/constants/queryKey";
 
 const useFollowList = (targetId: number, type: "follower" | "following") => {
-  const isFollower = type === "follower";
-  const queryKey = isFollower ? "following_id" : "random_id";
-
-  return useQuery({
-    queryKey: ["followListDetail", targetId, type],
-    queryFn: async () => {
-      const res = await axiosInstance.get("/follow_table", {
-        params: {
-          [`${queryKey}`]: `eq.${targetId}`,
-          is_following: "eq.true",
-          select: "random_id, following_id",
-        },
-      });
-      return res.data;
-    },
+  const {
+    data = [],
+    isLoading,
+    isError,
+  } = useQuery({
+    queryKey: [QUERY_KEYS.followListDetail, targetId, type],
+    queryFn: () => fetchFollowList(targetId, type),
   });
+
+  return {
+    followList: data,
+    isLoading,
+    isError,
+  };
 };
 
 export default useFollowList;
