+++ src/api/useFollowMutation.ts
@@ -1,32 +1,32 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import axiosInstance from "@/api/axiosInstance";
+import { postFollow, deleteFollow } from "@/db/follow";
 
 const useFollowMutation = () => {
   const queryClient = useQueryClient();
 
   const follow = useMutation({
-    mutationFn: (data: { fromId: number; toId: number }) =>
-      axiosInstance.post("/follow_table", {
-        random_id: data.fromId,
-        following_id: data.toId,
-        is_following: true,
-      }),
+    mutationFn: ({ fromId, toId }: { fromId: number; toId: number }) =>
+      postFollow({ fromId, toId }),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ["followCount"] });
     },
   });
 
   const unfollow = useMutation({
-    mutationFn: (id: number) =>
-      axiosInstance.patch(`/follow_table?id=eq.${id}`, {
-        is_following: false,
-      }),
+    mutationFn: (f_id: number) => deleteFollow(f_id),
     onSuccess: () => {
       queryClient.invalidateQueries({ queryKey: ["followCount"] });
     },
   });
 
-  return { follow, unfollow };
+  return {
+    follow: follow.mutate,
+    unfollow: unfollow.mutate,
+    isFollowPending: follow.isPending,
+    isUnfollowPending: unfollow.isPending,
+    isFollowError: follow.isError,
+    isUnfollowError: unfollow.isError,
+  };
 };
 
 export default useFollowMutation;
