+++ src/api/useLikedPlaylists.ts
@@ -2,15 +2,24 @@ import { useQuery } from "@tanstack/react-query";
 import { getAllPlaylistCardData } from "@/db/playlistCardData";
 import { getMyLikedPlaylistIds } from "@/db/like";
 import { useMemo } from "react";
+import { QUERY_KEYS } from "@/constants/queryKey";
 
 const useLikedPlaylists = (randomId?: number) => {
-  const { data: allPlaylists = [], ...restPlaylists } = useQuery({
-    queryKey: ["allPlaylists"],
+  const {
+    data: allPlaylists = [],
+    isLoading: isPlaylistsLoading,
+    isError: isPlaylistsError,
+  } = useQuery({
+    queryKey: [QUERY_KEYS.allPlaylists],
     queryFn: getAllPlaylistCardData,
   });
 
-  const { data: likedIds = [], ...restLikes } = useQuery({
-    queryKey: ["myLikedIds", randomId],
+  const {
+    data: likedIds = [],
+    isLoading: isLikesLoading,
+    isError: isLikesError,
+  } = useQuery({
+    queryKey: [QUERY_KEYS.myLikedIds, randomId],
     queryFn: () => getMyLikedPlaylistIds(randomId!),
     enabled: !!randomId,
   });
@@ -21,7 +30,10 @@ const useLikedPlaylists = (randomId?: number) => {
 
   return {
     likedPlaylists,
-    isLoading: restPlaylists.isLoading || restLikes.isLoading,
+    isLoading: isPlaylistsLoading || isLikesLoading,
+    isError: isPlaylistsError || isLikesError,
+    isPlaylistsError,
+    isLikesError,
   };
 };
