+++ src/api/useMyPlaylists.ts
@@ -1,12 +1,23 @@
 import { useQuery } from "@tanstack/react-query";
-import { getMyPlaylists } from "@/api/getMyPlaylists";
+import { getMyPlaylists } from "@/db/playlistCardData";
+import { QUERY_KEYS } from "@/constants/queryKey";
 
 const useMyPlaylists = (randomId?: number) => {
-  return useQuery({
-    queryKey: ["myPlaylists", randomId],
+  const {
+    data = [],
+    isLoading,
+    isError,
+  } = useQuery({
+    queryKey: [QUERY_KEYS.myPlaylists, randomId],
     queryFn: () => getMyPlaylists(randomId!),
     enabled: !!randomId,
   });
+
+  return {
+    myPlaylists: data,
+    isLoading,
+    isError,
+  };
 };
 
 export default useMyPlaylists;
