+++ src/api/useUpdateUserInfo.ts
@@ -1,9 +1,10 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { updateUser } from "@/db/users";
 import { useUserStore } from "@/stores/userStore";
+import { QUERY_KEYS } from "@/constants/queryKey";
 
 interface UpdateUserInfoParams {
-  id: number; // user_table의 id (PK)
+  id: number;
   updatedFields: {
     nickname?: string;
     sort_intro?: string;
@@ -11,23 +12,21 @@ interface UpdateUserInfoParams {
   };
 }
 
-// 유저 정보 수정 + 전역 상태 업데이트
 const useUpdateUserInfo = (onSuccessCallback?: () => void) => {
   const setUser = useUserStore((state) => state.setUser);
   const queryClient = useQueryClient();
 
-  return useMutation({
+  const { mutate, isPending, isError } = useMutation({
     mutationFn: ({ id, updatedFields }: UpdateUserInfoParams) =>
       updateUser(id, updatedFields),
     onSuccess: (updatedUser) => {
-      setUser(updatedUser); // 최신 정보로 상태 갱신
-      queryClient.invalidateQueries({ queryKey: ["currentUser"] });
+      setUser(updatedUser);
+      queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.currentUser] });
       onSuccessCallback?.();
     },
-    onError: (error) => {
-      console.error("유저 정보 업데이트 실패:", error.message);
-    },
   });
+
+  return { mutate, isPending, isError };
 };
 
 export default useUpdateUserInfo;
