+++ src/constants/queryKey.ts
@@ -1,0 +1,12 @@
+export const QUERY_KEYS = {
+  followerCount: "followerCount",
+  followingCount: "followingCount",
+  followListDetail: "followListDetail",
+  followStatus: "followStatus",
+  myLikedIds: "myLikedIds",
+  allPlaylists: "allPlaylists",
+  myPlaylists: "myPlaylists",
+  currentUser: "currentUser",
+  otherUser: "otherUser",
+  profileImage: "profileImage",
+};
