+++ src/db/comment.ts
@@ -82,13 +82,13 @@ export async function getCommentCountByPlaylist(
 }
 
 // 현재로서 사용하지 않음
-export async function patchComment(): Promise<Comment[]> {
-  const response = await axiosInstance.patch<Comment[]>("/comments_table");
-  return response.data;
-}
+// export async function patchComment(): Promise<Comment[]> {
+//   const response = await axiosInstance.patch<Comment[]>("/comments_table");
+//   return response.data;
+// }
 
 // 현재로서 사용하지 않음
-export async function deleteComment(): Promise<Comment[]> {
-  const response = await axiosInstance.delete<Comment[]>("/comments_table");
-  return response.data;
-}
+// export async function deleteComment(): Promise<Comment[]> {
+//   const response = await axiosInstance.delete<Comment[]>("/comments_table");
+//   return response.data;
+// }
