+++ src/db/like.ts
@@ -9,10 +9,10 @@ export interface Like {
 }
 
 //가져오기
-export async function getLike(): Promise<Like[]> {
-  const response = await axiosInstance.get<Like[]>("/likes_table");
-  return response.data;
-}
+// export async function getLike(): Promise<Like[]> {
+//   const response = await axiosInstance.get<Like[]>("/likes_table");
+//   return response.data;
+// }
 
 // 좋아요 여부 확인
 export async function getLikeStatus(
