+++ src/pages/play/Play.tsx
@@ -275,6 +275,7 @@ const ProfileImage = styled.img`
   width: 50px;
   height: 50px;
   border-radius: 50px;
+  object-fit: cover;
   cursor: pointer;
 `;
